+++ demo/src/main/java/net/vrallev/android/demo/cat/App.java
@@ -3,6 +3,8 @@ package net.vrallev.android.demo.cat;
 import android.app.Application;
 
 import net.vrallev.android.cat.CatGlobal;
+import net.vrallev.android.cat.instance.CatSimple;
+import net.vrallev.android.demo.cat.def.TestDefaultPackage;
 import net.vrallev.android.demo.cat.disabled.TestDisabled;
 
 /**
@@ -24,5 +26,7 @@ public class App extends Application {
 
         CatGlobal.setPackageEnabled(TestDisabled.class.getPackage().getName(), false);
         CatGlobal.setTagEnabled("Disabled", false);
+
+        CatGlobal.setDefaultCatLogPackage(TestDefaultPackage.class.getPackage().getName(), new CatSimple("DefaultPackage"));
     }
 }