+++ demo/src/main/java/net/vrallev/android/demo/cat/def/TestDefaultPackage.java
@@ -1,0 +1,18 @@
+package net.vrallev.android.demo.cat.def;
+
+import net.vrallev.android.cat.Cat;
+
+/**
+ * @author rwondratschek
+ */
+public final class TestDefaultPackage {
+
+    private TestDefaultPackage() {
+        // no op
+    }
+
+    public static void log() {
+        Cat.d("Hello from the default instance");
+    }
+
+}