+++ library/src/main/java/net/vrallev/android/cat/print/AndroidLog.java
@@ -1,5 +1,7 @@
 package net.vrallev.android.cat.print;
 
+import android.support.annotation.NonNull;
+import android.support.annotation.Nullable;
 import android.util.Log;
 
 /**
@@ -8,7 +10,11 @@ import android.util.Log;
 public class AndroidLog implements CatPrinter {
 
     @Override
-    public void println(int priority, String tag, String msg) {
-        Log.println(priority, tag, msg);
+    public void println(int priority, @NonNull String tag, @NonNull String message, @Nullable Throwable t) {
+        if (t != null) {
+            message += '\n' + Log.getStackTraceString(t);
+        }
+
+        Log.println(priority, tag, message);
     }
 }