+++ library/src/main/java/net/vrallev/android/cat/print/CatPrinter.java
@@ -1,9 +1,12 @@
 package net.vrallev.android.cat.print;
 
+import android.support.annotation.NonNull;
+import android.support.annotation.Nullable;
+
 /**
  * @author rwondratschek
  */
 public interface CatPrinter {
 
-    void println(int priority, String tag, String msg);
+    void println(int priority, @NonNull String tag, @NonNull String message, @Nullable Throwable t);
 }