+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 1.0.1 (2015-08-13)
+
+- bug fixes
+- update CatPrinter interface
+- add option to set a default CatLog for a specific package
+
 ## 1.0.0 (2015-08-12)
 
 - initial release